[058-cc7] Moved service and renamed state.
