+++ src/services/workspace-service.js
@@ -18,8 +18,22 @@ const getWorkspaces = async () => {
 	return [];
 };
 
+const postWorkspaces = async (workspaceData) => {
+
+    const response = await fetch(ENDPOINTS.WORKSPACE.USER_WORKSPACE, {
+		method: "POST",
+		body: JSON.stringify(workspaceData),
+	});
+
+	if (response.status !== 201) {
+		return false;
+	}
+	return true;
+};
+
 const workspaceService = {
 	getWorkspaces,
+	postWorkspaces,
 };
 
 export default workspaceService;
